+++ test/test_udp_socket.rs
@@ -7,7 +7,7 @@ use super::localhost;
 use std::io::net::ip::{Ipv4Addr};
 use mio::event as evt;
 
-type TestEventLoop = EventLoop<uint, ()>;
+type TestEventLoop = EventLoop<usize, ()>;
 
 const LISTENER: Token = Token(0);
 const SENDER: Token = Token(1);
@@ -32,7 +32,7 @@ impl UdpHandler {
     }
 }
 
-impl Handler<uint, ()> for UdpHandler {
+impl Handler<usize, ()> for UdpHandler {
     fn readable(&mut self, event_loop: &mut TestEventLoop, token: Token, _: evt::ReadHint) {
         match token {
             LISTENER => {
